refactor: rename response types file
