+++ src/lib/api_app/src/api/routes/mod.rs
@@ -1,5 +1,5 @@
 use super::{
-    routing_types::{APIRoutingResponse, APIRoutingError, ParsedRequest},
+    response_types::{APIRoutingResponse, APIRoutingError, ParsedRequest},
 };
 
 pub mod application;
